requirements.txt
PyQt5==5.15.10
openai==1.20.0
bcrypt==4.1.2
mysql-connector-python==8.3.0
python-docx==1.1.0